# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sankalpa Odisha API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /announcements/{id}/attachments:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      tags: [Announcements]
      operationId: listAnnouncementAttachments
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AttachmentListResponse" } } } }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
    post:
      tags: [Announcements]
      operationId: registerAnnouncementAttachment
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RegisterAttachmentRequest" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/AttachmentResponse" } } } }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
  /attachments/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    delete:
      tags: [Announcements]
      operationId: deleteAttachment
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
  /storage/uoi-notes/uploads/request-url:
    post:
      operationId: requestUoiNoteUploadUrl
      tags: [storage]
      summary: Request an upload URL for a UOI note attachment
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RequestUoiNoteUploadUrlBody" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/RequestUploadUrlResponse" } } } }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
  /uoi-notes/{id}/attachments:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      tags: [UoiNotes]
      operationId: listUoiNoteAttachments
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/UoiNoteAttachmentListResponse" } } } }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
    post:
      tags: [UoiNotes]
      operationId: registerUoiNoteAttachment
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RegisterAttachmentRequest" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/UoiNoteAttachmentResponse" } } } }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /uoi-note-attachments/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    delete:
      tags: [UoiNotes]
      operationId: deleteUoiNoteAttachment
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
...
  /uoi-notes:
    get:
      tags: [uoi-notes]
      operationId: listUoiNotes
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: pageSize
          schema: { type: integer }
        - in: query
          name: departmentId
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: all
          description: When true, returns every matching note without pagination (used for downloads).
          schema: { type: boolean }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/UoiNoteListResponse" } } } }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
    post:
      tags: [uoi-notes]
      operationId: createUoiNote
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUoiNoteRequest" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/UoiNoteResponse" } } } }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
...
  /uoi-notes/{id}:
    put:
      tags: [uoi-notes]
      operationId: updateUoiNote
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUoiNoteRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/UoiNoteResponse" } } } }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
    delete:
      tags: [uoi-notes]
      operationId: deleteUoiNote
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: No Content }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
components:
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    AttachmentRecord:
      type: object
      properties:
        id: { type: integer }
        announcementId: { type: integer }
        objectPath: { type: string }
        fileName: { type: string }
        fileSize: { type: integer }
        contentType: { type: string }
        uploadedBy: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, announcementId, objectPath, fileName, fileSize, contentType, uploadedBy, createdAt]
    AttachmentListResponse:
      type: object
      properties:
        attachments:
          type: array
          items: { $ref: "#/components/schemas/AttachmentRecord" }
      required: [attachments]
    AttachmentResponse:
      type: object
      properties:
        attachment: { $ref: "#/components/schemas/AttachmentRecord" }
      required: [attachment]
    RegisterAttachmentRequest:
      type: object
      properties:
        objectPath: { type: string }
        fileName: { type: string }
        fileSize: { type: integer }
        contentType: { type: string }
        uploadToken:
          type: string
          description: >
            Server-issued one-time token from POST /storage/uploads/request-url
            that binds the upload to this actor + objectPath. Required.
      required: [objectPath, fileName, fileSize, contentType, uploadToken]
...
    RequestUoiNoteUploadUrlBody:
      type: object
      properties:
        name: { type: string }
        size: { type: integer }
        contentType: { type: string }
        uoiNoteId:
          type: integer
          description: >
            UOI note this upload is intended for. Bound into the issued
            uploadToken so the file cannot be registered against a
            different note.
      required: [name, size, contentType, uoiNoteId]
    UoiNoteAttachmentRecord:
      type: object
      properties:
        id: { type: integer }
        noteId: { type: integer }
        objectPath: { type: string }
        fileName: { type: string }
        fileSize: { type: integer }
        contentType: { type: string }
        uploadedBy: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, noteId, objectPath, fileName, fileSize, contentType, uploadedBy, createdAt]
    UoiNoteAttachmentListResponse:
      type: object
      properties:
        attachments:
          type: array
          items: { $ref: "#/components/schemas/UoiNoteAttachmentRecord" }
      required: [attachments]
    UoiNoteAttachmentResponse:
      type: object
      properties:
        attachment: { $ref: "#/components/schemas/UoiNoteAttachmentRecord" }
      required: [attachment]
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
        uploadToken:
          type: string
          description: >
            Short-lived HMAC token bound to {actor, objectPath, expiry}.
            Must be replayed when calling POST /announcements/{id}/attachments.
        metadata:
          type: object
          properties:
            name: { type: string }
            size: { type: integer }
            contentType: { type: string }
          required: [name, size, contentType]
      required: [uploadURL, objectPath, uploadToken, metadata]
...
    UoiNote:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        date: { type: string }
        departmentId: { type: integer }
        departmentName: { type: string }
        occasion: { type: string, nullable: true }
        location: { type: string, nullable: true }
        description: { type: string }
        createdBy: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, title, date, departmentId, departmentName, description, createdBy, createdAt]
    UoiNoteResponse:
      type: object
      properties:
        note: { $ref: "#/components/schemas/UoiNote" }
      required: [note]
    UoiNoteListResponse:
      type: object
      properties:
        notes:
          type: array
          items: { $ref: "#/components/schemas/UoiNote" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
      required: [notes, total, page, pageSize]
    CreateUoiNoteRequest:
      type: object
      properties:
        title: { type: string }
        date: { type: string }
        departmentId: { type: integer }
        occasion: { type: string }
        location: { type: string }
        description: { type: string }
      required: [title, date, departmentId, description]
...